+++ RayTracer/src/Samplers/RegularSampler.cs
@@ -48,7 +48,7 @@ namespace RayTracer
                 {
                     for(int k = 0; k < n; k++)
                     {
-                        samples.Add(new Point2D((float)k / (float)n, (float)j / (float)n));
+                        samples.Add(new Point2D((double)k / (double)n, (double)j / (double)n));
                     }
                 }
             }
